+++ src/helpers.c
@@ -1,6 +1,6 @@
 #include "helpers.h"
 
-Point2 CastPoint2(double x, double y) {
+Point2 CastPoint2(float x, float y) {
 	Point2 newPoint;
 	newPoint.x = x;
 	newPoint.y = y;
@@ -8,7 +8,7 @@ Point2 CastPoint2(double x, double y) {
 	return newPoint;
 }
 
-Point3 CastPoint3(double x, double y, double z) {
+Point3 CastPoint3(float x, float y, float z) {
 	Point3 newPoint;
 	newPoint.x = x;
 	newPoint.y = y;
